+++ KeyManager.Library.UI/SymmetricKeyGenerationDialog.xaml.cs
@@ -77,7 +77,7 @@ namespace Leosac.KeyManager.Library.UI
 
         private void BtnRandom_Click(object sender, RoutedEventArgs e)
         {
-            KeyValue = KeyGeneration.Random(KeySize > 0 ? KeySize : 16);
+            KeyValue = KeyGeneration.Random((uint)(KeySize > 0 ? KeySize : 16));
         }
 
         private void BtnPassword_Click(object sender, RoutedEventArgs e)
